+++ domain/src/main/kotlin/com/patrykandpatryk/liftapp/domain/unit/UnitConverter.kt
@@ -2,13 +2,13 @@ package com.patrykandpatryk.liftapp.domain.unit
 
 interface UnitConverter {
 
-    suspend fun convertToPreferredUnit(from: LongDistanceUnit, value: Float): Float
+    suspend fun convertToPreferredUnit(from: LongDistanceUnit, value: Double): Double
 
-    suspend fun convertToPreferredUnit(from: MediumDistanceUnit, value: Float): Float
+    suspend fun convertToPreferredUnit(from: MediumDistanceUnit, value: Double): Double
 
-    suspend fun convertToPreferredUnit(from: ShortDistanceUnit, value: Float): Float
+    suspend fun convertToPreferredUnit(from: ShortDistanceUnit, value: Double): Double
 
-    suspend fun convertToPreferredUnit(from: MassUnit, value: Float): Float
+    suspend fun convertToPreferredUnit(from: MassUnit, value: Double): Double
 
-    suspend fun convertToPreferredUnitAndFormat(from: ValueUnit, vararg values: Float): String
+    suspend fun convertToPreferredUnitAndFormat(from: ValueUnit, vararg values: Double): String
 }
